More 64-bit fixes
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@9321 a1c6a512-1295-4272-9138-f99709370657

+++ apps/plugins/doom/d_main.c
@@ -806,7 +806,7 @@ void D_DoomMainSetup(void)
       // for statistics driver
       extern  void* statcopy;
 
-      statcopy = (void*)atoi(myargv[p+1]);
+      statcopy = (void*)(long)atoi(myargv[p+1]);
       printf ("External statistics registered.\n");
    }
